+++ src/conversion.rs
@@ -44,7 +44,7 @@ where
     #[inline]
     fn as_ptr(&self) -> *mut ffi::PyObject {
         self.as_ref()
-            .map_or_else(std::ptr::null_mut, |t| t.into_ptr())
+            .map_or_else(std::ptr::null_mut, |t| t.as_ptr())
     }
 }
 
@@ -555,7 +555,7 @@ where
 #[cfg(test)]
 mod tests {
     use crate::types::{IntoPyDict, PyAny, PyDict, PyList};
-    use crate::{Python, ToPyObject};
+    use crate::{AsPyPointer, PyObject, Python, ToPyObject};
 
     use super::PyTryFrom;
 
@@ -593,6 +593,23 @@ mod tests {
             let list = PyList::new(py, &[1, 2, 3]);
             let val = unsafe { <PyList as PyTryFrom>::try_from_unchecked(list.as_ref()) };
             assert_eq!(list, val);
+        });
+    }
+
+    #[test]
+    fn test_option_as_ptr() {
+        Python::with_gil(|py| {
+            let mut option: Option<PyObject> = None;
+            assert_eq!(option.as_ptr(), std::ptr::null_mut());
+
+            let none = py.None();
+            option = Some(none.clone());
+
+            let ref_cnt = none.get_refcnt(py);
+            assert_eq!(option.as_ptr(), none.as_ptr());
+
+            // Ensure ref count not changed by as_ptr call
+            assert_eq!(none.get_refcnt(py), ref_cnt);
         });
     }
 }